Moving Card component to new folder
