+++ packages/ui/index.tsx
@@ -1,2 +1,2 @@
 // component exports
-export { Card } from "./card";
+export { Card } from "./components/card";
